Development of the Landing page, Set-up of Redux framework, and Routing completed.
In this PR, I have completed the initial development of the landing page.
I have also set-up the redux framework for the application, which will ease the
inter-component communication within the application architechture.
I have also added the relevant routes in the 'App.jsx' file, and enabled routing between them.