Added Routing and Redux to the application
This PR incorporates routing and redux in the application, and a brief functional demo of
the code is presented in the 'LandingPage' component, that showcases the expected usage.
The code being submitted has been written using the latest recommendations in mind,
like using react hooks for Routing and redux, as well as conforming to ESLint.